requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
openai
python-dotenv
pydantic
